pyproject.toml
[project]
name = "whisper-bot"
version = "0.1.0"
description = "Telegram bot for voice message transcription using Whisper"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.0",
    "python-dotenv>=1.0",
    "httpx>=0.24",
    "whisper-transcribe @ file:///Users/alexeyprudkin/dev/whisper-cli",
]

[project.scripts]
whisper-bot = "whisper_bot.bot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/whisper_bot"]
